[Refactor] [GGFE-178] modalProvider 리팩토링
