+++ components/admin/feedback/FeedbackTable.tsx
@@ -29,9 +29,9 @@ const tableTitle: { [key: string]: string } = {
 export interface IFeedback {
   id: number;
   intraId: string;
-  category: number; // 1: bug, 2: suggestion, 3: question
-  content: string;
-  createdAt: Date;
+  category?: number; // 1: bug, 2: suggestion, 3: question
+  content?: string;
+  createdAt?: Date;
   isSolved: boolean;
 }
